# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "Core_me",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "permissions": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "person": {
                      "additionalProperties": true,
                      "nullable": true,
                      "properties": {
                        "firstName": {
                          "type": "string"
                        },
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "jobTitle": {
                          "nullable": true,
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "firstName",
                        "lastName",
                        "jobTitle"
                      ],
                      "type": "object"
                    },
                    "user": {
                      "additionalProperties": false,
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "roles": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      },
                      "required": [
                        "id",
                        "roles"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "user",
                    "person",
                    "permissions"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Principal corrente, persona collegata e permessi effettivi",
        "tags": [
          "core"
        ]
      }
    },
...
    "/api/v1/org-units": {
      "get": {
        "operationId": "Core_listOrgUnits",
        "parameters": [
          {
            "in": "query",
            "name": "tree",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      },
      "post": {
        "operationId": "Core_createOrgUnit",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "code": {
                    "maxLength": 40,
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 120,
                    "minLength": 1,
                    "type": "string"
                  },
                  "parentId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
    "/api/v1/org-units/{id}": {
      "delete": {
        "operationId": "Core_archiveOrgUnit",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      },
      "patch": {
        "operationId": "Core_updateOrgUnit",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "code": {
                    "maxLength": 40,
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 120,
                    "minLength": 1,
                    "type": "string"
                  },
                  "parentId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
    "/api/v1/people": {
      "get": {
        "operationId": "Core_listPeople",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "orgUnitId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "managerId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "invited",
                "active",
                "leaving",
                "terminated",
                "suspended"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      },
      "post": {
        "operationId": "Core_createPerson",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "customFields": {
                    "additionalProperties": {},
                    "type": "object"
                  },
                  "email": {
                    "format": "email",
                    "type": "string"
                  },
                  "employeeNumber": {
                    "maxLength": 50,
                    "type": "string"
                  },
                  "firstName": {
                    "maxLength": 100,
                    "minLength": 1,
                    "type": "string"
                  },
                  "hireDate": {
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "type": "string"
                  },
                  "jobLevel": {
                    "maxLength": 60,
                    "type": "string"
                  },
                  "jobTitle": {
                    "maxLength": 120,
                    "type": "string"
                  },
                  "lastName": {
                    "maxLength": 100,
                    "minLength": 1,
                    "type": "string"
                  },
                  "location": {
                    "maxLength": 120,
                    "type": "string"
                  },
                  "managerId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "orgUnitId": {
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "firstName",
                  "lastName"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
    "/api/v1/people/{id}": {
      "get": {
        "operationId": "Core_getPerson",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      },
      "patch": {
        "operationId": "Core_updatePerson",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "customFields": {
                    "additionalProperties": {},
                    "type": "object"
                  },
                  "email": {
                    "format": "email",
                    "type": "string"
                  },
                  "employeeNumber": {
                    "maxLength": 50,
                    "type": "string"
                  },
                  "firstName": {
                    "maxLength": 100,
                    "minLength": 1,
                    "type": "string"
                  },
                  "hireDate": {
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "type": "string"
                  },
                  "jobLevel": {
                    "maxLength": 60,
                    "type": "string"
                  },
                  "jobTitle": {
                    "maxLength": 120,
                    "type": "string"
                  },
                  "lastName": {
                    "maxLength": 100,
                    "minLength": 1,
                    "type": "string"
                  },
                  "location": {
                    "maxLength": 120,
                    "type": "string"
                  },
                  "managerId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "orgUnitId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "status": {
                    "enum": [
                      "active",
                      "leaving",
                      "suspended"
                    ],
                    "type": "string"
                  },
                  "terminationDate": {
                    "nullable": true,
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
    "/api/v1/people/{id}/history": {
      "get": {
        "operationId": "Core_personHistory",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
    "/api/v1/people/{id}/terminate": {
      "post": {
        "operationId": "Core_terminate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
    "/api/v1/people/import": {
      "post": {
        "operationId": "Core_importPeople",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "createOrgUnits": {
                    "default": false,
                    "type": "boolean"
                  },
                  "csv": {
                    "maxLength": 5000000,
                    "minLength": 1,
                    "type": "string"
                  },
                  "dryRun": {
                    "default": true,
                    "type": "boolean"
                  }
                },
                "required": [
                  "csv"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Import persone da CSV. dryRun=true restituisce anteprima ed errori senza scrivere (CORE-012).",
        "tags": [
          "core"
        ]
      }
    },
    "/api/v1/people/import/template": {
      "get": {
        "operationId": "Core_importTemplate",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
...
    "/api/v1/role-assignments": {
      "post": {
        "operationId": "Core_assignRole",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "role": {
                    "enum": [
                      "tenant_admin",
                      "hr_admin",
                      "hrbp",
                      "manager",
                      "employee",
                      "observer",
                      "analyst"
                    ],
                    "type": "string"
                  },
                  "scopeId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "scopeType": {
                    "default": "tenant",
                    "enum": [
                      "tenant",
                      "org_unit"
                    ],
                    "type": "string"
                  },
                  "userId": {
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "userId",
                  "role"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
    "/api/v1/role-assignments/{id}": {
      "delete": {
        "operationId": "Core_revokeRole",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
...
    "/api/v1/tenant": {
      "get": {
        "operationId": "Core_tenant",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      },
      "patch": {
        "operationId": "Core_updateTenant",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "defaultLocale": {
                    "enum": [
                      "it",
                      "en"
                    ],
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 120,
                    "minLength": 1,
                    "type": "string"
                  },
                  "settings": {
                    "additionalProperties": {},
                    "type": "object"
                  },
                  "timezone": {
                    "maxLength": 60,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
...
    "/api/v1/users/{id}/disable": {
      "post": {
        "operationId": "Core_disable",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "core"
        ]
      }
    },
...
  },
  "info": {
    "title": "WorkingBetter API",
    "description": "API REST unica per web, mobile e connettori (ADR-0005). Errori in formato RFC 9457 (application/problem+json). I body e le query sono descritti dagli stessi schemi Zod usati per la validazione (ADR-0009).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "description": "URI del tipo di errore (about:blank per gli errori generici)"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "description": "Identificativo della richiesta (x-request-id)"
          },
          "code": {
            "type": "string",
            "description": "Codice applicativo stabile (ErrorCodes in @wb/shared)"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "path",
                "message"
              ],
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}